feat/menambahkan fitur untuk edit video youtube

+++ app/Http/Controllers/BlogController.php
@@ -9,10 +9,9 @@ use Google\Service\YouTube\VideoStatus;
 use App\Models\Post;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
-use Google\Http\MediaFileUpload;
 use Google\Service\YouTube\Video;
 use Illuminate\Support\Facades\Auth;
-
+use Laravel\Socialite\Facades\Socialite;
 
 class BlogController extends Controller
 {
@@ -40,71 +39,6 @@ class BlogController extends Controller
     {
         return view('admin.blogs.create');
     }
-
-    /**
-     * YOUTUBE UPLOAD
-     */
-    // public function uploadVideoToYotube($videoPath, $judul, $deskripsi) {
-    //     $user = Auth::user();
-    //     $googleToken = $user->google_token;
-    //     $googleRefreshToken = $user->google_refresh_token;
-
-    //     $client = new Client();
-    //     $client->setClientId(env('GOOGLE_CLIENT_ID'));
-    //     $client->setClientSecret(env('GOOGLE_CLIENT_SECRET'));
-    //     $client->setRedirectUri(env('GOOGLE_REDIRECT_URI'));
-    //     $client->addScope(YouTube::YOUTUBE_UPLOAD);
-    //     $client->setAccessToken([
-    //         'access_token' => $googleToken,
-    //         'expires_in' => 10800, // 3jam
-    //     ]);
-
-    //      // Periksa apakah token sudah kedaluwarsa
-    //     if ($client->isAccessTokenExpired()) {
-    //         // Gunakan refresh token untuk mendapatkan token baru
-    //         $newToken = $client->fetchAccessTokenWithRefreshToken($googleRefreshToken);
-            
-    //         // Perbarui token di database
-    //         $user->update([
-    //             'google_token' => $newToken['access_token'],
-    //         ]);
-
-    //         $client->setAccessToken($newToken);
-    //     }
-
-    //     $youtube = new YouTube($client);
-    //     $video = new Video();
-    //     $snippet = new VideoSnippet();
-    //     $status = new VideoStatus();
-         
-    //     $snippet->setTitle($judul);
-    //     $snippet->setDescription($deskripsi);
-         
-    //     $status->setPrivacyStatus('private');
-
-    //     $video->setSnippet($snippet);
-    //     $video->setStatus($status);
-
-    //     $chunkSizeBytes = 1 * 1024 * 1024; // 1 MB
-    //     $client->setDefer(true);
-
-    //     $insertRequest = $youtube->videos->insert('snippet,status', $video);
-    //     $media = new MediaFileUpload($client, $insertRequest, 'video/*', null, true, $chunkSizeBytes);
-    //     $media->setFileSize(filesize($videoPath));
-
-    //     // Upload video dalam chunk
-    //     $handle = fopen($videoPath, 'rb');
-    //         while (!feof($handle)) {
-    //         $chunk = fread($handle, $chunkSizeBytes);
-    //         $status = $media->nextChunk($chunk);
-    //     }
-    //     fclose($handle);
-
-    //     $client->setDefer(false);
-
-    //     return $status; // Mengembalikan respons dari YouTube
-    // }
-    
 
     /**
      * Store a newly created resource in storage.
@@ -149,53 +83,25 @@ class BlogController extends Controller
             ]
         );
 
-        // upload gambar
+        # upload gambar
         if($request->hasFile('thumbnail')){
             $image = $request->file('thumbnail');
             $image_name = time() . '_' . $image->getClientOriginalName();
             $path_location = public_path(getenv('THUMBNAILS_LOCATION'));
-            // $path_location = base_path('../' . env('THUMBNAILS_LOCATION', 'upload/thumbnails')); // untuk host infinityfree
+            # untuk host infinityfree
+            // $path_location = base_path('../' . env('THUMBNAILS_LOCATION', 'upload/thumbnails'));
             $image->move($path_location, $image_name);
         };
 
-        $dataStore = [
-            'judul'=> $request->judul,
-            'slug'=> $this->generateSlug($request->judul),
-            'deskripsi'=> $request->deskripsi,
-            'konten'=> $request->konten,
-            'kategori'=> $request->kategori,
-            'status'=> $request->status,
-            'thumbnail' => isset($image_name) ? $image_name : null,
-            'user_id' => $user->id
-        ];
-
-        $post = Post::create($dataStore);
-
+        # upload video
         if ($request->hasFile('video_url')) {
             $video = $request->file('video_url');
-            // $video_name = $video->getClientOriginalName();
             $video_name = time() . '_' . $video->getClientOriginalName();
             $video_path_location = public_path('upload/videos');
+            # untuk di infinityfree
+            // $video_path_location = base_path('../' . 'upload/videos');
             $video->move($video_path_location, $video_name);
 
-            // Periksa apakah token sudah kedaluwarsa
-            // if ($client->isAccessTokenExpired()) {
-            //     // Gunakan refresh token untuk mendapatkan token baru
-            //     $newToken = $client->fetchAccessTokenWithRefreshToken($googleRefreshToken);
-                
-            //     // Perbarui token di database
-            //     $user->update([
-            //         'google_token' => $newToken['access_token'],
-            //     ]);
-
-            //     $client->setAccessToken($newToken);
-            // } else {
-            //     $client->setAccessToken([
-            //         'access_token' => $googleToken,
-            //         'expires_in' => 7200, // 2 jam
-            //     ]);
-            // }
-            
             $service = new Youtube($client);
             $snippet = new VideoSnippet();
             $status = new VideoStatus();
@@ -203,13 +109,13 @@ class BlogController extends Controller
 
             $snippet->setTitle($request->judul);
             $snippet->setDescription(Str::of($request->konten)->stripTags());
-            $status->setPrivacyStatus('private');
+            $status->setPrivacyStatus('unlisted');
             $video_youtube->setSnippet($snippet);
             $video_youtube->setStatus($status);
 
             $videoFile = public_path('upload/videos' . '/' . $video_name);
-
-            // dd($videoFile);
+            # untuk di infinityfree
+            // $videoFile = base_path('../upload/videos' . '/' . $video_name);
 
             $response = $service->videos->insert(
                 'snippet,status',
@@ -220,11 +126,21 @@ class BlogController extends Controller
                     'uploadType' => 'multipart'
                 )
             );
-
-            // Menyimpan ID video YouTube ke dalam post
-            $post->youtube_video_id = $response->getId(); // Menyimpan ID video YouTube
-            $post->save();
         }
+
+        $dataStore = [
+            'judul'=> $request->judul,
+            'slug'=> $this->generateSlug($request->judul),
+            'deskripsi'=> $request->deskripsi,
+            'konten'=> $request->konten,
+            'kategori'=> $request->kategori,
+            'status'=> $request->status,
+            'thumbnail' => isset($image_name) ? $image_name : null,
+            'user_id' => $user->id,
+            'youtube_video_id' => $response->getId()
+        ];
+
+        Post::create($dataStore);
         
         return redirect()->route('admin.blogs.index')->with('success', 'Data berhasil ditambahkan');
     }
@@ -251,13 +167,33 @@ class BlogController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $apiKey = env('GOOGLE_API_KEY');
+        $user = Auth::user();
+        $youtubeId = $post->youtube_video_id;
+        $googleToken = $user->google_token;
+        
+        $client = new Client();
+        $client->setAuthConfig(base_path('google-oauth-client.json'));
+        $client->setDeveloperKey($apiKey);
+        $client->setRedirectUri(env('GOOGLE_REDIRECT_URI'));
+        $client->addScope(Youtube::YOUTUBE, Youtube::YOUTUBE_FORCE_SSL, Youtube::YOUTUBE_CHANNEL_MEMBERSHIPS_CREATOR);
+        $client->setAccessType('offline');
+        $client->setApprovalPrompt('force');
+        
+        $client->setAccessToken([
+            'access_token' => $googleToken,
+        ]);
+
+        $service = new Youtube($client);
+        $snippet = new VideoSnippet();
+        $video_youtube_data = new Video();
+        
         $request->validate(
             [
                 'judul' => 'required',
                 'konten' => 'required',
                 'thumbnail' => 'image|mimes:jpeg,jpg,png,JPG,JPEG|max:10240',
-                'kategori' => 'required',
-
+                'video' => 'file|mimetypes:video/mp4,video/avi,video/mkv|max:51200'
             ],
             [
                 'judul.required'=> 'Judul wajib diisi',
@@ -265,17 +201,18 @@ class BlogController extends Controller
                 'thumbnail.image'=> 'Hanya gambar yang diperbolehkan',
                 'thumbnail.mimes'=> 'Format gambar hanya JPEG, JPG dan PNG',
                 'thumbnail.max'=> 'Size maksimum 10MB',
-                'kategori.required'=> 'Wajib memilih kategori',
+                'video.mimetypes' => 'Format video hanya MP4, AVI, atau MKV',
+                'video.max' => 'Ukuran maksimum video adalah 50MB',
             ]
         );
 
         // update gambar
-        if($request->hasFile('thumbnail')){
+        if($post->kategori == 'artikel' && $request->hasFile('thumbnail')){
             if(isset($post->thumbnail) && file_exists(public_path(getenv('THUMBNAILS_LOCATION')) . '/' . $post->thumbnail)) {
                 unlink(public_path(getenv('THUMBNAILS_LOCATION')) . '/' . $post->thumbnail);
             }
 
-            // untuk host infinityfree
+            # untuk host infinityfree
             // if(isset($post->thumbnail) && file_exists(base_path('../' . env('THUMBNAILS_LOCATION', 'upload/thumbnails')) . '/' . $post->thumbnail)) {
             //     unlink(base_path('../' . env('THUMBNAILS_LOCATION', 'upload/thumbnails')) . '/' . $post->thumbnail);
             // }
@@ -283,15 +220,34 @@ class BlogController extends Controller
             $image = $request->file('thumbnail');
             $image_name = time() . '_' . $image->getClientOriginalName();
             $path_location = base_path('../' . getenv('THUMBNAILS_LOCATION'));
+            # untuk hosting di infinityfree
+            // $path_location = public_path(env('THUMBNAILS_LOCATION')); 
             $image->move($path_location, $image_name);
         };
 
-         $dataUpdate = [
+        if($post->kategori == 'video') {
+            // Mendapatkan detail snippet untuk pra-edit
+            $response = $service->videos->listVideos(
+                'snippet',
+                ['id' => $youtubeId]
+            );
+            
+            $video = $response[0];
+            $snippet = $video->snippet;
+
+            $video_youtube_data->setId($youtubeId);
+            $snippet->setTitle($request->judul);
+            $snippet->setDescription(Str::of($request->konten)->stripTags());
+            $video_youtube_data->setSnippet($snippet);  
+            
+            $response = $service->videos->update('snippet', $video_youtube_data);
+        }
+
+        $dataUpdate = [
             'judul'=> $request->judul,
             'slug'=> $this->generateSlug($request->judul, $post->id),
             'deskripsi'=> $request->deskripsi,
             'konten'=> $request->konten,
-            'kategori'=> $request->kategori,
             'status'=> $request->status,
             'thumbnail' => isset($image_name) ? $image_name : $post->thumbnail
         ];
